Config

+++ src/keleqram_bot.cpp
@@ -1,18 +1,14 @@
 #include "keleqram_bot.hpp"
-#include "request.hpp"
 #include <chrono>
-#include <sstream>
 
 namespace keleqram {
-static TimePoint         initial_time = std::chrono::system_clock::now();
-static const char*       START_COMMAND  {"start"};
-static const char*       HELP_COMMAND   {"help"};
-static const char*       DELETE_COMMAND {"delete"};
-static const char*       TOKEN          {""};
-static const char*       DEFAULT_REPLY  {"Defeat Global Fascism"};
-static const char*       DEFAULT_RETORT {"I hear you, bitch"};
-static const char*       MARKDOWN_MODE  {"Markdown"};
-static const uint32_t    FIFTH_OF_DAY   {17280};
+static const char*       START_COMMAND   {"start"};
+static const char*       HELP_COMMAND    {"help"};
+static const char*       DELETE_COMMAND  {"delete"};
+static const char*       MESSAGE_COMMAND {"message"};
+static const char*       TOKEN           {""};
+static const char*       DEFAULT_REPLY   {"Defeat Global Fascism"};
+static const char*       MARKDOWN_MODE   {"Markdown"};
 static const uint32_t    KANYE_URL_INDEX   {0};
 static const uint32_t    ZENQUOTE_URL_INDEX{1};
 static const uint32_t    BTC_URL_INDEX     {2};
@@ -33,209 +29,60 @@ static const char*       URLS[] {
 static const int64_t     CHAT_IDs[] {
 
 };
+static const int32_t     ADMIN_IDs[] {
 
-const int64_t DEFAULT_CHAT_ID = *(CHAT_IDs);
-/**
- * Global var
- */
-static kint8_t chat_idx{};
+};
+static const int32_t     ADMIN_NUM{0x01};
+       const int64_t     DEFAULT_CHAT_ID = *(CHAT_IDs);
+static kint8_t           chat_idx{};
 
 /**
   ┌──────────────────────────────────────────────────────────┐
   │░░░░░░░░░░░░░░░░░░░░░░░░░░ Helpers ░░░░░░░░░░░░░░░░░░░░░░░│
-  └──────────────────────────────────────────────────────────┘
-*/
-
+  └──────────────────────────────────────────────────────────┘**/
 static void LogMessage(const MessagePtr& message)
 {
   log(std::string{"User "                  }, message->from->firstName,
+      std::string{" with UID "             }, std::to_string(message->from->id),
       std::string{" from chat "            }, message->chat->title,
       std::string{" with ID "              }, std::to_string(message->chat->id),
       std::string{" said the following: \n"}, message->text);
 }
 
-static bool ActionTimer(uint32_t duration = FIFTH_OF_DAY)
+/**
+ * IsAdmin
+ */
+static bool IsAdmin(const int32_t& id)
 {
-  const TimePoint now = std::chrono::system_clock::now();
-  const int64_t   elapsed = std::chrono::duration_cast<Duration>(now - initial_time).count();
-
-  if (elapsed > duration)
-  {
-    initial_time = now;
-    return true;
-  }
-
+  for (auto i = 0; i < ADMIN_NUM; i++)
+    if (ADMIN_IDs[i] == id)
+      return true;
   return false;
 }
 
-static std::string FloatToDecimalString(float n)
-{
-  std::stringstream ss;
-  ss << std::fixed << std::setprecision(2) << n;
-  return ss.str();
-}
-
-static std::string ToLower(const std::string& s)
-{
-  std::string t{s};
-  std::transform(t.begin(), t.end(), t.begin(), [](char c) { return tolower(c); });
-  return t;
-}
-
-static void Hello(TgBot::Bot& bot)
-{
-  printf("Bot username: %s\n", bot.getApi().getMe()->username.c_str());
-}
-
-static std::string DecodeHTML(const std::string& text)
-{
-  std::string                                  decoded{};
-  std::unordered_map<std::string, std::string> convert({
-    {"&quot;",  "\""},
-    {"&apos;",  "'"},
-    {"&amp;",   "&"},
-    {"&gt;",    ">"},
-    {"&lt;",    "<"},
-    {"&frasl;", "/"}});
-
-  for (size_t i = 0; i < text.size(); ++i)
-  {
-    bool flag = false;
-    for (const auto& [key, value] : convert)
-    {
-      if (i + key.size() - 1 < text.size())
-      {
-        if (text.substr(i, key.size()) == key)
-        {
-          decoded += value;
-          i += key.size() - 1;
-          flag = true;
-          break;
-        }
-      }
-    }
-
-  if (!flag)
-    decoded += text[i];
-
-  }
-
-  return decoded;
-}
-
-struct MimeType
-{
-  std::string name;
-  bool        video;
-
-  bool IsVideo() const { return video;    }
-  bool IsPhoto() const { return !(video); }
-};
-
-static MimeType GetMimeType(const std::string& path)
-{
-  const auto it = path.find_last_of('.');
-  if (it != std::string::npos)
-  {
-    const auto extension = ToLower(path.substr(it + 1));
-
-    if (extension == "jpg" || extension == "jpeg")
-      return MimeType{"image/jpeg", false};
-    else
-    if (extension == "png")
-      return MimeType{"image/png", false};
-    else
-    if (extension == "gif")
-      return MimeType{"image/gif", false};
-    else
-    if (extension == "mp4")
-      return MimeType{"video/mp4", true};
-    else
-    if (extension == "mkv")
-      return MimeType{"video/mkv", true};
-    else
-    if (extension == "webm")
-      return MimeType{"video/webm", true};
-    else
-    if (extension == "mpeg")
-      return MimeType{"video/mpeg", true};
-    else
-    if (extension == "mov")
-      return MimeType{"video/quicktime", true};
-  }
-  return MimeType{"unknown", false};
-}
-
-static std::string ExtractTempFilename(const std::string& full_url)
-{
-  static const char* TEMP_FILE{"temp_file"};
-
-        auto ext_end = full_url.find_first_of('?');
-        ext_end      = ext_end == std::string::npos ? full_url.size() : ext_end;
-  const auto url     = full_url.substr(0, ext_end);
-  const auto ext_beg = url.find_last_of('.');
-  const auto ext_len = (ext_beg != url.npos) ? (url.size() - ext_beg) : 0;
-  const auto filename = (ext_len > 0) ? TEMP_FILE + url.substr(ext_beg, ext_len) : TEMP_FILE;
-  return filename;
-}
-
-static std::string FetchTemporaryFile(const std::string& full_url, const bool verify_ssl = true)
-{
-  const auto filename   = ExtractTempFilename(full_url);
-  const cpr::Response r = cpr::Get(cpr::Url{full_url}, cpr::VerifySsl(verify_ssl));
-  SaveToFile(r.text, filename);
-
-  return filename;
-}
-
 /**
- * @brief Chunk Message
+ * Greeting
  *
- * @param   [in]  {std::string} message
- * @returns [out] {std::vector<std::string>}
+ * Bot greeting and list of commands
  */
-static std::vector<std::string> const ChunkMessage(const std::string& message) {
-  static const uint32_t MAX_CHUNK_SIZE = TELEGRAM_CHAR_LIMIT - 6;
-
-  std::vector<std::string>     chunks{};
-  const std::string::size_type message_size {message.size()};
-  const std::string::size_type num_of_chunks{message.size() / MAX_CHUNK_SIZE + 1};
-  uint32_t                     chunk_index  {1};
-  std::string::size_type       bytes_chunked{ };
-
-  if (num_of_chunks > 1)
-  {
-    chunks.reserve(num_of_chunks);
-    while (bytes_chunked < message_size)
-    {
-      const std::string::size_type size_to_chunk =
-        (bytes_chunked + MAX_CHUNK_SIZE > message_size) ?
-          (message_size - bytes_chunked) :
-          MAX_CHUNK_SIZE;
-
-      std::string oversized_chunk = message.substr(bytes_chunked, size_to_chunk);
-
-      const std::string::size_type ws_idx = oversized_chunk.find_last_of(" ") + 1;
-      const std::string::size_type pd_idx = oversized_chunk.find_last_of(".") + 1;
-      const std::string::size_type index  = (size_to_chunk > MAX_CHUNK_SIZE) ?
-         (ws_idx > pd_idx) ?  ws_idx :
-                              pd_idx
-                           :
-         size_to_chunk;
-
-      chunks.emplace_back((!index) ?
-        oversized_chunk :
-        oversized_chunk.substr(0, index) + '\n' +
-        std::to_string(chunk_index++)    + '/'  + std::to_string(num_of_chunks));
-
-      bytes_chunked += index;
-    }
-  }
-  else
-    chunks.emplace_back(message);
-
-  return chunks;
-}
+static std::string Greeting(MessagePtr& message)
+{
+  static const char* BotInfo{
+    "**Available commands:**\n```\n"
+    "/kanye        - Timeless advice and inspiration\n"
+    "/quote        - Quotes from persons that are not Kanye West\n"
+    "/btc          - Latest BTC price\n"
+    "/link         - Latest LINK price\n"
+    "/eth          - Latest Ethereum price\n"
+    "/insult       - You deserve what you get\n"
+    "/wiki <query> - Search Wikipedia```"};
+  const auto& name = (message->newChatMember) ?
+                      (message->newChatMember->firstName.empty()) ?
+                        message->from->firstName : message->newChatMember->firstName :
+                     (message->from) ? message->from->firstName : "the room";
+  const auto& room = message->chat->title;
+  return "Welcome to " + room + ", " + name + "\n\n" + BotInfo;
+};
 
 /**
  * GetRequest
@@ -296,31 +143,6 @@ static std::string GetRequest(uint32_t url_index)
 
   return text;
 }
-
-/**
- *
- */
-static std::string ExtractWikiText(const nlohmann::json& json)
-{
-  std::string text{};
-  if (!json.is_null() && json.is_object() && json.contains("query"))
-  {
-    for (const auto& item : json["query"]["search"])
-    {
-      std::string s = item["snippet"].get<std::string>();
-      for (auto it = s.find("</span>"); it != std::string::npos;)
-      {
-        s   = s.substr(it + 7);
-        it  = s.find("</span>");
-      }
-      text += s;
-    }
-  }
-
-  auto decoded_text = DecodeHTML(text);
-  return decoded_text;
-}
-
 /**
  * GetWiki
  */
@@ -344,31 +166,6 @@ static std::string GetWiki(std::string message)
   }
   return text;
 }
-
-/**
- * Greeting
- *
- * Bot greeting and list of commands
- */
-static std::string Greeting(MessagePtr& message)
-{
-  static const char* BotInfo{
-    "**Available commands:**\n```\n"
-    "/kanye        - Timeless advice and inspiration\n"
-    "/quote        - Quotes from persons that are not Kanye West\n"
-    "/btc          - Latest BTC price\n"
-    "/link         - Latest LINK price\n"
-    "/eth          - Latest Ethereum price\n"
-    "/insult       - You deserve what you get\n"
-    "/wiki <query> - Search Wikipedia```"};
-  const auto& name = (message->newChatMember) ?
-                      (message->newChatMember->firstName.empty()) ?
-                        message->from->firstName : message->newChatMember->firstName :
-                     (message->from) ? message->from->firstName : "the room";
-  const auto& room = message->chat->title;
-  return "Welcome to " + room + ", " + name + "\n\n" + BotInfo;
-};
-
 /**
  * HandleRequest
  * @static
@@ -415,10 +212,15 @@ KeleqramBot::KeleqramBot(const std::string& token)
   tx(0),
   rx(0),
   tx_err(0),
-  rx_err(0)
+  rx_err(0),
+  tx_msgs(TXMessages{})
 {
+  if (config.ParseError())
+    throw std::invalid_argument{"Unable to load config"};
+
   Hello(m_bot);
   SetListeners();
+  SetReplies(m_replies);
 }
 
 /**
@@ -432,19 +234,52 @@ void KeleqramBot::Poll()
 }
 
 /**
+ * HandlePrivateRequest
+ */
+void KeleqramBot::HandlePrivateRequest(MessagePtr message)
+{
+  const auto GetRequestInfo = [](const std::string& s) -> RequestInfo
+  {
+    static const size_t prefix_num = 9;
+                 auto   rem        = s.substr(prefix_num);
+           const auto   idx        = rem.find_first_of(' ');
+           const auto   uid        = rem.substr(0, idx - 1);
+           const auto   msg        = rem.substr(idx + 1);
+           const auto   id         = (IsAllNum(uid)) ? uid : "@channel" + uid;
+    return RequestInfo{id, msg};
+  };
+
+  try
+  {
+    if (!IsAdmin(message->from->id)) return;
+
+    const auto info = GetRequestInfo(message->text);
+    SendMessage(info.message, info.id);
+  }
+  catch (const std::exception& e)
+  {
+    log("Failed to parse private message request. Exception: ", e.what());
+  }
+}
+
+/**
  * SetListeners
  */
 void KeleqramBot::SetListeners()
 {
-  m_bot.getEvents().onCommand   (START_COMMAND,  [this](MessagePtr message)
+  m_bot.getEvents().onCommand   (START_COMMAND,   [this](MessagePtr message)
   {
     SendMessage("Hi!", message->chat->id);
   });
-  m_bot.getEvents().onCommand   (HELP_COMMAND,   [this](MessagePtr message)
+  m_bot.getEvents().onCommand   (HELP_COMMAND,    [this](MessagePtr message)
   {
     SendMessage(Greeting(message), message->chat->id, MARKDOWN_MODE);
   });
-  m_bot.getEvents().onCommand   (DELETE_COMMAND, [this](MessagePtr message)
+  m_bot.getEvents().onCommand   (MESSAGE_COMMAND, [this](MessagePtr message)
+  {
+    HandlePrivateRequest(message);
+  });
+  m_bot.getEvents().onCommand   (DELETE_COMMAND,  [this](MessagePtr message)
   {
     DeleteMessages(message);
   });
@@ -542,7 +377,7 @@ void KeleqramBot::HandleMessage(MessagePtr message)
   LogMessage(message);
 
   if (reply_message && IsReply(id, reply_message->messageId))
-    SendMessage(DEFAULT_RETORT, id);
+    SendMessage(m_replies.at(GetRandom(0, m_replies.size())), id);
   else
   if (IsEvent(message))
     HandleEvent(message);
@@ -567,31 +402,27 @@ void KeleqramBot::HandleEvent(MessagePtr message)
 void KeleqramBot::DeleteMessages(MessagePtr message)
 {
   using ChatMsgs = std::vector<int32_t>;
-  const auto GetNum = [](const std::string& s) -> int32_t
+  const int64_t     chat_id  = message->chat->id;
+  const auto        action   = DeleteAction(message->text);
+        ChatMsgs&   messages = tx_msgs[chat_id];
+  if (action.valid && messages.size())
   {
-                 int32_t n{};
-    static const char*   prefix{"/delete last "};
-    static const size_t  prefix_length{13};
-           const size_t  idx = s.find(prefix);
-    if (!idx)
-    {
-      const auto& rem = s.substr(prefix_length);
-      if (isdigit(rem.front())) // TODO: check them all, bitch
-        n = std::stoi(rem);
-    }
-    return n;
-  };
+    auto it = (messages.end() - action.n);
 
-  const int64_t     chat_id    = message->chat->id;
-        ChatMsgs&   messages   = tx_msgs[chat_id];
+    m_api.deleteMessage(chat_id, *(it));
+    it = messages.erase(it);
 
-  if (messages.size())
-    for (auto it = (messages.end() - GetNum(message->text)); it != messages.end();)
+    if (action.delete_last)
     {
-      m_api.deleteMessage(chat_id, *(it));
-      it = messages.erase(it);
+      while (it != messages.end())
+      {
+        m_api.deleteMessage(chat_id, *(it));
+        it = messages.erase(it);
+      }
     }
+  }
 }
+
 /**
   ┌──────────────────────────────────────────────────────────┐
   │░░░░░░░░░░░░░░░░░░░░░░ Specializations ░░░░░░░░░░░░░░░░░░░░│
